pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shuttersort"
version = "0.1.0"
description = "AI-powered media folder analyzer and pruner using local Vision models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "camiloavilacm", email = "camiloavilacm@users.noreply.github.com"},
]
keywords = ["media", "cleanup", "ollama", "vision", "ai", "photos", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "ollama>=0.1.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
    "rawpy>=0.18.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
shuttersort = "media_pruner.cli:main"

[project.urls]
Homepage = "https://github.com/camiloavilacm/ShutterSort"
Repository = "https://github.com/camiloavilacm/ShutterSort"
Issues = "https://github.com/camiloavilacm/ShutterSort/issues"

[tool.setuptools.packages.find]
include = ["media_pruner*"]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["media_pruner", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort (import sorting)
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["media_pruner"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require a real Ollama instance (deselect with '-m \"not integration\"')",
]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "rawpy",
    "cv2",
    "ollama",
]
ignore_missing_imports = true
